This commit reflects a major revison in the code base following extensive testing outside of the repository
While the new code base is not fully functioning yet it marks the beginning of the transition to a multi-protcol python based
internal script/payload. Data collection in the simmulation has also been slightly revised and further evaulaition of that awaits
the intial completation of the python version of the ssh connection system.

This iteration of the code can form a ssh tunnel, but some connection and runnign commands needs to be revised and improved. Other
issues may exists. This commit is beign made because the code has reached a point where it starting to resemble its intended
purpose as well as for accountability and devlopment protability reasons. The way networks and ip assingment is handled has been
modified as well; this does work as intended.

In short, major "of-sceen" twaeks were done to beggin the trasintion for a new model for the simmulator. It does not work yet, but
enough does that it is getting moved to dev to keep the project up to date.

As a final note not everything is python and the python code itself uses system calls so it is still bound to Linux